Remove "pull requests" document
I'm not sure that this is still valuable. Removing for now in the
interest of keeping things minimal and current, but happy for pushback
here.

If we do end up keeping this, I think we should label (both in the title
and body) that this info is specific to Github.