Add 'Git Development Workflow' document
This describes at a high-level the process of using git to record
and share changes. It's a pretty good overview of the entire flow,
and it depends on later documents going into more detail on some of
the points.

There was some discussion among the team about whether to include
the git commands or not. I thought they would be a little
distracting from the flow of the overview. But Caleb mentioned that
we also have (for example) Academy folks learning from these
documents, so the commands are useful to include. We settled on what
is, I think, a good compromise: include the commands as an appendix.
This makes it easy for those familiar with basic git commands to
skip, while still being present for those less familiar.