Add a project README best practice
Most of this content is thanks to Tim Irwin, who created a README
template in Q2 2016 for a craftsmanship team rock. The associated
Trello card is here:
https://trello.com/c/D1mvDfO8/365-create-templates-for-a-readme-plus-other-internal-documentation-and-score

The README template is copied verbatim (modulo Markdown quoting)
from that document, which is available in Google Drive here:
https://docs.google.com/document/d/1_l-33_o3SB8y2gj4dZPcHvU3HjYl3dSwTjUe_ahrn64/edit?ts=59b6b4ee

I moved the "Additional Documentation" section from the document,
including a paragraph I added about architecture decision records,
out of the template and into a separate section.

I also added the other text preceding the template.